fixes warnings when building for 64-bit Linux

+++ gameplay/src/Model.cpp
@@ -308,8 +308,8 @@ static bool drawWireframe(Mesh* mesh)
 
 static bool drawWireframe(MeshPart* part)
 {
-    unsigned int indexCount = part->getIndexCount();
-    unsigned int indexSize = 0;
+    uintptr_t indexCount = part->getIndexCount();
+    uintptr_t indexSize = 0;
     switch (part->getIndexFormat())
     {
     case Mesh::INDEX8:
@@ -330,7 +330,7 @@ static bool drawWireframe(MeshPart* part)
     {
     case Mesh::TRIANGLES:
         {
-            for (unsigned int i = 0; i < indexCount; i += 3)
+            for (uintptr_t i = 0; i < indexCount; i += 3)
             {
                 GL_ASSERT( glDrawElements(GL_LINE_LOOP, 3, part->getIndexFormat(), ((const GLvoid*)(i*indexSize))) );
             }
@@ -339,7 +339,7 @@ static bool drawWireframe(MeshPart* part)
 
     case Mesh::TRIANGLE_STRIP:
         {
-            for (unsigned int i = 2; i < indexCount; ++i)
+            for (uintptr_t i = 2; i < indexCount; ++i)
             {
                 GL_ASSERT( glDrawElements(GL_LINE_LOOP, 3, part->getIndexFormat(), ((const GLvoid*)((i-2)*indexSize))) );
             }
